+++ apaveinput.tcl
@@ -34,7 +34,7 @@
 
 package require Tk
 
-package provide apave 3.3a3
+package provide apave 3.3
 
 source [file join [file dirname [info script]] apavedialog.tcl]
 
@@ -304,7 +304,7 @@ oo::class create ::apave::APaveInput {
         -t 1 -head "\nAPaveInput returned an error: \n" -hfg red -weight bold
       return 0
     }
-    if {[lindex $res 0]!=1} {  ;# restore old values if OK not chosen
+    if {![lindex $res 0]} {  ;# restore old values if OK not chosen
       foreach {vn vv} $_savedvv {
         # tk_optionCascade (destroyed now) was tracing its variable => catch
         catch {set $vn $vv}